!Hotfix: read에 실패하면 main page를 로드할 수 없음. read 요청 이슈 해결 중

See merge request s08-bigdata-recom-sub2/S08P22A503!207

Former-commit-id: ee2af55655f2581b3578a0e31c166ff0e959bee6

+++ backend/src/main/java/com/a503/churros/service/news/NewsService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface NewsService {
     List<Integer> sendRecommend(long userId);
     List<Integer> sendSample();
-    void saveReadArti(long userId, int articleId);
+    void saveReadArti(long userId, Long articleId);
     void recordLike(long userId, long articleId);
     List<Long> getLikeList(long userIdx);
     void recordDisLike(long userId, long articleId);
