+++ backend/src/main/java/com/a503/churros/service/news/NewsServiceImpl.java
@@ -43,7 +43,7 @@ public class NewsServiceImpl implements NewsService{
     }
 
     @Override
-    public void saveReadArti(long userId, int articleId) {
+    public void saveReadArti(long userId, Long articleId) {
         Read read = rr.findByUserIdxAndArticleIdx(userId , articleId).orElse(null);
         if(read == null){
             read = Read.builder()
